Add the Dockerfile used to generate the image
Reenable clang-tidy on new pipeline

Fix vera and clang-tidy

Update the used Dockerfile to include clang-tidy

Clean up commented out stuff from CI configuration

Update Dockerfile according to suggestions

Fix the virtual package

Move update to initialization to ensure latest version of everything

Merge branch 'faster-ci' of https://gitlab.com/acubesat/obc/ecss-services into faster-ci

Fixes for new broken clang-tidy rules

+++ src/Services/EventReportService.cpp
@@ -120,7 +120,7 @@ void EventReportService::listOfDisabledEventsReport() {
 
 	uint16_t numberOfDisabledEvents = stateOfEvents.size() - stateOfEvents.count();
 	report.appendHalfword(numberOfDisabledEvents);
-	for (uint16_t i = 0; i < stateOfEvents.size(); i++) {
+	for (size_t i = 0; i < stateOfEvents.size(); i++) {
 		if (not stateOfEvents[i]) {
 			report.appendEnum16(i);
 		}
